Separação do controlador web da classe principal
